move `TimerLabelStyle` to its own file.
