+++ LifeCoachWatchOS/Timer Feature/UI/TimerView.swift
@@ -1,20 +1,6 @@
 import SwiftUI
 import Combine
 
-struct TimerLabelStyle: LabelStyle {
-    let isLuminanceReduced: Bool
-    let customFont: String?
-    
-  func makeBody(configuration: Configuration) -> some View {
-    configuration
-          .title
-          .foregroundColor(.red)
-          .font(Font.custom(customFont ?? "", size: 40))
-          .opacity(isLuminanceReduced ? 0.75 : 1.0)
-  }
-}
-
-
 public struct TimerView: View {
     @Environment(\.isLuminanceReduced) var isLuminanceReduced
